refactor: update ProductCard import path for placeholders utility and add new placeholder functions for better organization and type clarity
