+++ app/components/ProductCard.tsx
@@ -5,7 +5,7 @@ import type {MoneyV2, Product} from '@shopify/hydrogen/storefront-api-types';
 import {AddToCartButton} from '~/components/AddToCartButton';
 import {Link} from '~/components/Link';
 import {Text} from '~/components/Text';
-import {getProductPlaceholder} from '~/lib/placeholders';
+import {getProductPlaceholder} from '~/lib/utils/placeholders';
 import {isDiscounted, isNewArrival} from '~/lib/utils';
 import clsx from 'clsx';
 import type {ProductCardFragment} from 'storefrontapi.generated';
