+++ app/lib/utils/placeholders.ts
